pyproject.toml
[project]
name = "typer-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.70.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "typer>=0.24.1",
]
